pyproject.toml
[project]
name = "cryptography-labs"
version = "0.1.0"
description = "Cryptography labs"
readme = "README.md"
requires-python = ">=3.14.0, <3.15"
dependencies = [
    "gostcrypto==1.2.5",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "mimesis==19.1.0",
    "pandas>=3.0.1",
    "pandoc==2.4",
    "playwright==1.58.0",
    "pycryptodome==3.23.0",
    "sympy==1.14.0",
    "uv-build>=0.9.2",
    "wonderwords==3.0.1",
]

[build-system]
requires = ["uv_build>=0.9.2"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "labs"
module-root = ""

[tool.uv]
#required-version = "~=0.9.2"
add-bounds = "exact"
